Use templates instead of static files to avoid a puppet bug.
I'm not quite sure what the bug is, but it recently started happening!
This is a huge problem, because it interrupts the puppet run, and it
doesn't complete in one shot. This is a big fail for vagrant
environments. It might be due to a new puppet update! Who knows!